requirements.txt
# ETL Dependencies for DW_Salud Project
# Python 3.13.5

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Database Connectivity
pyodbc>=5.0.0
sqlalchemy>=2.0.0

# Configuration and Environment
python-dotenv>=1.0.0
